pyproject.toml
[project]
name = "parallel-execution-rf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "robotframework>=7.1.1",
]
